blackman window

+++ Jack_AntiLarsen/src/iir.cpp
@@ -32,8 +32,8 @@ preFiltersBank::preFiltersBank(float f_sampling_in, float gb_in, float q_factor_
     damp = sqrtf(std::abs(1.0f - powf(gb_in, 2.0f))) / gb_in;
     for (auto &filter: this->filters) {
         wo = 2.0f * PI * (f) / f_sampling_in;
-        filter.e = 1.0f/(1.0f + damp*tan(wo/(q_factor_in * 2.0f)));
-        filter.p = cos(wo);
+        filter.e = 1.0f/(1.0f + damp*tanf(wo/(q_factor_in * 2.0f)));
+        filter.p = cosf(wo);
         filter.d[0] = filter.e;
         filter.d[1] = 2*filter.e*filter.p;
         filter.d[2] = (2*filter.e-1);
